pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "peon"
version = "0.1.0"
description = "zug zug"
requires-python = ">=3.12"

dependencies = ["python-dotenv", "pyyaml", "psutil", "aiohttp"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/peon"]
